Changes and improvements to pixel/screen handler
- Changed the way drawing pixels is done, now handled by screen handler instead of pixel handler.

- Changed the way pixels are drawn, now done using fillRect instead of changing canvasdata.
This speeds up drawing considerably, and actually makes this possible to run.

- Added a PixelType and Id to Pixel.
PixelType allows for different functionality based on pixeltype (sand vs water), while the ID allows me to compare pixels for the celauto function.

- Added the cellulair automata functionality to PixelHandler.
Currently, every pixel compares itself to all other pixels in the pixel array. If the x and y are the same, it will push itself up by reverse pixel falling speed, causing it to stack on eachother.